datasetgate/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datasetgate"
version = "0.1.0"
description = "Unified authorization service for neuroscience datasets"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.1,<6.0",
    "djangorestframework>=3.15,<4.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-cloud-storage>=2.0",
    "gunicorn>=22.0",
    "scim2-filter-parser>=0.5",
    "markdown>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "datasetgate.settings"
python_files = ["tests.py", "test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
